Moe of unix in sys/wait and sys/types.

+++ libc/common.c
@@ -22,15 +22,28 @@ int execvpe(const char *file, char *const argv[], char *const envp[])
     return __syscall3(EXECVPE, (uint64_t)file, (uint64_t)argv, (uint64_t)envp);
 }
 
-pid_t wait(uint64_t *status)
+//pid_t wait(uint64_t *status)
+//{
+//    return __syscall1(WAIT, (uint64_t)status);
+//}
+
+//pid_t waitpid(pid_t pid, uint64_t *status, int options)
+//{
+//    return __syscall3(WAITPID, (uint64_t)pid, (uint64_t)status, (uint64_t)options);
+//}
+
+pid_t wait(int* status)
 {
-    return __syscall1(WAIT, (uint64_t)status);
+	return __syscall1(WAIT, (uint64_t)status);
 }
 
-pid_t waitpid(pid_t pid, uint64_t *status, int options)
+pid_t waitpid(pid_t pid, int* status, int options)
 {
-    return __syscall3(WAITPID, (uint64_t)pid, (uint64_t)status, (uint64_t)options);
+	return __syscall3(WAITPID, (uint64_t)pid, 
+		(uint64_t)status, (uint64_t)options);
 }
+
+
 
 void listprocess()
 {
